requirements.txt
pytest
appium-python-client
requests
python-dotenv